Problem 20: Part 1
Part 2 partially complete.  We an get corners without actually building
the map by just calculating the neighbor-set and returning the 4 pieces
that have exactly 2 neighbors.  Pieces that have 3 neighbors will be on
a non-corner edge, and pieces with 4 neighbors will be "inside".

To do part 2, we need to actually assemble the map, which involves not
only calculating neighbors, but fixing the orientations in place.  This
requires keeping track of which permutations were used.  The tileMap
data structure would need to be a nested map to keep track of which
edges match per permutation.